execcmd manager init/start task: Decoupled Linux and Windows code files
