+++ agent/engine/execcmd/manager_windows.go
@@ -1,0 +1,27 @@
+// +build windows
+
+// Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"). You may
+// not use this file except in compliance with the License. A copy of the
+// License is located at
+//
+//	http://aws.amazon.com/apache2.0/
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package execcmd
+
+import (
+	"path/filepath"
+
+	"github.com/aws/amazon-ecs-agent/agent/config"
+)
+
+var (
+	hostExecDepsDir = filepath.Join(config.AmazonECSProgramFiles, "managed-agents", "execute-command")
+	HostBinDir      = filepath.Join(hostExecDepsDir, "bin")
+)
